Hack to uncomment updateParserOutput() again
Requires EntitySchema commit “Hack to support StatementDataUpdater”
(currently dd9e9723c5 but will probably be rebased a bunch of times).
The ReferencedEntitiesDataUpdater hack should later become a hook, but
my current feeling is that the hook should happen here, and that we want
to keep EntityTitleLookup, a relatively core interface, free of
pseudo-entities. (But this may turn out to be infeasible.)

Change-Id: I8c3d9bd3d727a77997778c3232eadc68262058e6