Hack Special:EntityPage to not crash on E1
Looking up the Title for an EntitySchema ID is not supported. Exit early
with a “bad request / inalid ID” error, as if the ID couldn’t be parsed
at all.

I wonder whether hooking EntitySchemaId parsing into the EntityIdParser
service is unwise, and we should do it at another level closer to the
surface instead, so that it works when parsing a data value, but not
everywhere else.

Change-Id: I32178b0d2c647c85eebe8d409f7b846446134b9a