Introduce HackPseudoEntityIdParser service
And remove EntitySchemaId support from the general EntityIdParser;
instead, the new HackPseudoEntityIdParser wraps the general
EntityIdParser, adding EntitySchemaId support, but is used only in the
specific places where we need it. General-purpose entity ID parsing is
thus free from pseudo-entities.

Remains to be seen how this works out compared to the previous approach,
I think? But to me it seems safer to “lift” the EntitySchemaId support
closer to the “surface” and keep it out of the core data model.

Change-Id: I2763f29659fcc5b4ff317fbb488076b311de2612