Introduce PseudoEntityId and IndeterminateEntityId
With the decision to make EntitySchema not a proper Wikibase Entity type
using the Entity Registration, while still returning it with a
datavalue-type of `wikibase-entityid`, we need to add some additional
extension points to Wikibase especially around parsing and deserializing
entity-id string literals. However, when we parse an EntitySchema id
as an EntityId, then it cannot be used to load an EntityDocument. Thus,
there are many places where we need to distinguish with pseudo-EntityIds
is, like EntitySchema, and the usual vanilla Wikibase EntityIds like
Items, Properties and Lexemes.

To have confidence when dealing with code about what kind of EntityId
we're dealing with, this change introduces two new interfaces.
The existing interface `EntityId` remains technically unchanged, even
though it now extends and has all its methods moved into a new upstream
interface `IndeterminateEntityId`.
The interface `IndeterminateEntityId` represents an EntityId that could
be either a usual vanilla Wikibase EntityId, or an EntityId that cannot
be used for a lot of things, especially not to load an actual Entity: a
`PseudoEntityId`.
The interface `PseudoEntityId` also extends `IndeterminateEntityId`, so
that `EntityId` and `PseudoEntityId` share the same methods.

However, this change only introduces the new interfaces and so has no
bahavior changes at all yet.

Note: I'd be very happy about a better name than IndeterminateEntityId

Change-Id: I52e7ef30b684d7828c3a57c34569db7f2d5dde73