Do not fail on parser functions
Parser functions try to load the Entity to show its label and add usage
tracking for that. However, loading PseudoEntities is not possible.

Note that this in particular does not add usage tracking for the
EntitySchema label. So if the label is ever made available later, all
pages using that will need to be purged separately.

This change also makes the listing of used entities on &action=edit and
&action=info not fail.

Change-Id: I1da024c1f93cc590808acb443f752c10f30fe299