Use PseudoEntityIdParser in EntityIdValue and deserializing
This, together with the respective changes in EntitySchema, is doing the
actual core work for allowing values in statements to be returned as
`wikibase-entityid` value type without being real Wikibase Entities in
code.

That means that with this change, EntityIdValue no longer always
contains an EntityId which can be used to look up an Entity.
EntityIdValue can now also contain PseudoEntityId which will need
special handling.

The following changes deal with the fallout from that and additional
functionality.

Change-Id: Ie0f1cd54b5f00b5a9a3f6db5f6a953fa51b77495